refactor: token validate Root로 위임 후 hook으로 분리
